Class ver 1.0
Class in C # supports a queue of limited size (20 elements). Each
element added to the queue should have a certain value (an integer in
the range 1-10). From the queue can be obtained element with the highest
value. In the case of overflow queue (adding 21 element) will be removed
from the queue element with the smallest value. Reading the item deletes
the downloaded item from the queue.